requirements.txt
asgiref==3.5.0
Django==3.2.10
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
